ft_atol update

+++ push_swap/include/libft/ft_atol.c
@@ -12,11 +12,11 @@
 
 #include "libft.h"
 
-long	ft_atol(const char *str)
+long long	ft_atol(const char *str)
 {
-	long	result;
-	int		sign;
-	int		i;
+	long long	result;
+	int			sign;
+	int			i;
 
 	if (!str)
 		return (0);
